- Move allChildren from beta to GA
