+++ EasyPost.Tests/ServicesTests/WithParameters/UserServiceTest.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using EasyPost.Models.API;
+using EasyPost.Parameters.User;
 using EasyPost.Tests._Utilities;
 using EasyPost.Tests._Utilities.Attributes;
 using EasyPost.Utilities.Internal.Attributes;
@@ -47,6 +48,27 @@ namespace EasyPost.Tests.ServicesTests.WithParameters
             Assert.IsType<User>(user);
             Assert.StartsWith("user_", user.Id);
             Assert.Equal("Test User", user.Name);
+        }
+
+        [Fact]
+        [CrudOperations.Read]
+        [Testing.Function]
+        public async Task TestAllChildren()
+        {
+            UseVCR("all_children");
+
+            Dictionary<string, object> fixture = new Dictionary<string, object> { { "page_size", Fixtures.PageSize } };
+
+            AllChildren parameters = Fixtures.Parameters.Users.AllChildren(fixture);
+
+            ChildUserCollection childUserCollection = await Client.User.AllChildren(parameters);
+            List<User> children = childUserCollection.Children;
+
+            Assert.True(children.Count <= Fixtures.PageSize);
+            foreach (User item in children)
+            {
+                Assert.IsType<User>(item);
+            }
         }
 
         [Fact]
