+++ EasyPost.Tests/cassettes/net/user_service/all_children.json
@@ -1,7 +1,7 @@
 [
   {
-    "Duration": 780,
-    "RecordedAt": "2023-12-12T16:48:04.664161-07:00",
+    "Duration": 368,
+    "RecordedAt": "2024-01-04T12:10:17.16264-07:00",
     "Request": {
       "Body": "",
       "BodyContentType": "Text",
@@ -11,7 +11,7 @@
         "Authorization": "<REDACTED>",
         "User-Agent": "<REDACTED>"
       },
-      "Uri": "https://api.easypost.com/beta/users/children?page_size=5"
+      "Uri": "https://api.easypost.com/v2/users/children?page_size=5"
     },
     "Response": {
       "Body": "{\"children\":[],\"has_more\":false}",
@@ -24,20 +24,19 @@
       "HttpVersion": "1.1",
       "ResponseHeaders": {
         "X-Frame-Options": "SAMEORIGIN",
-        "x-xss-protection": "1; mode=block",
+        "X-XSS-Protection": "1; mode=block",
         "X-Content-Type-Options": "nosniff",
         "x-download-options": "noopen",
         "x-permitted-cross-domain-policies": "none",
-        "referrer-policy": "strict-origin-when-cross-origin",
-        "x-ep-request-uuid": "a51ac9656578f134e7897f6e00535a50",
+        "Referrer-Policy": "strict-origin-when-cross-origin",
+        "x-ep-request-uuid": "ec7851ea65970299f43f18f3000ba2e9",
         "Cache-Control": "no-store, no-cache, private",
         "Pragma": "no-cache",
-        "x-runtime": "0.037266",
-        "Vary": "Origin",
-        "x-node": "bigweb41nuq",
-        "x-version-label": "easypost-202312122047-25e60d20e0-master",
+        "x-runtime": "0.055720",
+        "x-node": "bigweb33nuq",
+        "x-version-label": "easypost-202401041812-437974c716-master",
         "x-backend": "easypost",
-        "x-proxied": "intlb1nuq b3de2c47ef,extlb2nuq 003ad9bca0",
+        "x-proxied": "intlb1nuq 2c48984abf,extlb1nuq 003ad9bca0",
         "Strict-Transport-Security": "max-age=31536000; includeSubDomains; preload"
       },
       "Status": {
