+++ EasyPost.Tests/cassettes/net/user_service/get_next_page_of_children.json
@@ -1,7 +1,7 @@
 [
   {
-    "Duration": 404,
-    "RecordedAt": "2023-12-12T16:48:01.114217-07:00",
+    "Duration": 234,
+    "RecordedAt": "2024-01-04T12:10:17.44144-07:00",
     "Request": {
       "Body": "",
       "BodyContentType": "Text",
@@ -11,7 +11,7 @@
         "Authorization": "<REDACTED>",
         "User-Agent": "<REDACTED>"
       },
-      "Uri": "https://api.easypost.com/beta/users/children?page_size=5"
+      "Uri": "https://api.easypost.com/v2/users/children?page_size=5"
     },
     "Response": {
       "Body": "{\"children\":[],\"has_more\":false}",
@@ -29,15 +29,14 @@
         "x-download-options": "noopen",
         "x-permitted-cross-domain-policies": "none",
         "Referrer-Policy": "strict-origin-when-cross-origin",
-        "x-ep-request-uuid": "826f63d16578f131e7897f680054f891",
+        "x-ep-request-uuid": "ec7851ea65970299f01fd453000ba347",
         "Cache-Control": "no-store, no-cache, private",
         "Pragma": "no-cache",
-        "x-runtime": "0.055031",
-        "Vary": "Origin",
-        "x-node": "bigweb39nuq",
-        "x-version-label": "easypost-202312122047-25e60d20e0-master",
+        "x-runtime": "0.035129",
+        "x-node": "bigweb36nuq",
+        "x-version-label": "easypost-202401041812-437974c716-master",
         "x-backend": "easypost",
-        "x-proxied": "intlb1nuq b3de2c47ef,extlb1nuq 003ad9bca0",
+        "x-proxied": "intlb2nuq 2c48984abf,extlb1nuq 003ad9bca0",
         "Strict-Transport-Security": "max-age=31536000; includeSubDomains; preload"
       },
       "Status": {
