+++ EasyPost.Tests/cassettes/netstandard/user_service/all_children.json
@@ -1,7 +1,7 @@
 [
   {
-    "Duration": 754,
-    "RecordedAt": "2023-12-12T16:48:01.114217-07:00",
+    "Duration": 367,
+    "RecordedAt": "2024-01-04T12:10:28.956797-07:00",
     "Request": {
       "Body": "",
       "BodyContentType": "Text",
@@ -11,7 +11,7 @@
         "Authorization": "<REDACTED>",
         "User-Agent": "<REDACTED>"
       },
-      "Uri": "https://api.easypost.com/beta/users/children?page_size=5"
+      "Uri": "https://api.easypost.com/v2/users/children?page_size=5"
     },
     "Response": {
       "Body": "{\"children\":[],\"has_more\":false}",
@@ -24,20 +24,19 @@
       "HttpVersion": "1.1",
       "ResponseHeaders": {
         "X-Frame-Options": "SAMEORIGIN",
-        "X-XSS-Protection": "1; mode=block",
+        "x-xss-protection": "1; mode=block",
         "X-Content-Type-Options": "nosniff",
         "x-download-options": "noopen",
         "x-permitted-cross-domain-policies": "none",
-        "Referrer-Policy": "strict-origin-when-cross-origin",
-        "x-ep-request-uuid": "826f63d76578f131e7897f670054f892",
+        "referrer-policy": "strict-origin-when-cross-origin",
+        "x-ep-request-uuid": "3c78c409659702a4f451ac05000b147d",
         "Cache-Control": "no-store, no-cache, private",
         "Pragma": "no-cache",
-        "x-runtime": "0.041771",
-        "Vary": "Origin",
-        "x-node": "bigweb38nuq",
-        "x-version-label": "easypost-202312122047-25e60d20e0-master",
+        "x-runtime": "0.035621",
+        "x-node": "bigweb40nuq",
+        "x-version-label": "easypost-202401041812-437974c716-master",
         "x-backend": "easypost",
-        "x-proxied": "intlb2nuq b3de2c47ef,extlb1nuq 003ad9bca0",
+        "x-proxied": "intlb2nuq 2c48984abf,extlb2nuq 003ad9bca0",
         "Strict-Transport-Security": "max-age=31536000; includeSubDomains; preload"
       },
       "Status": {
