+++ EasyPost.Tests/cassettes/netstandard/user_service/get_next_page_of_children.json
@@ -1,7 +1,7 @@
 [
   {
-    "Duration": 284,
-    "RecordedAt": "2023-12-12T16:48:01.433485-07:00",
+    "Duration": 381,
+    "RecordedAt": "2024-01-04T12:10:29.405428-07:00",
     "Request": {
       "Body": "",
       "BodyContentType": "Text",
@@ -11,7 +11,7 @@
         "Authorization": "<REDACTED>",
         "User-Agent": "<REDACTED>"
       },
-      "Uri": "https://api.easypost.com/beta/users/children?page_size=5"
+      "Uri": "https://api.easypost.com/v2/users/children?page_size=5"
     },
     "Response": {
       "Body": "{\"children\":[],\"has_more\":false}",
@@ -24,20 +24,20 @@
       "HttpVersion": "1.1",
       "ResponseHeaders": {
         "X-Frame-Options": "SAMEORIGIN",
-        "X-XSS-Protection": "1; mode=block",
+        "x-xss-protection": "1; mode=block",
         "X-Content-Type-Options": "nosniff",
         "x-download-options": "noopen",
         "x-permitted-cross-domain-policies": "none",
-        "Referrer-Policy": "strict-origin-when-cross-origin",
-        "x-ep-request-uuid": "826f63d46578f131e7897f6a0054f8d2",
+        "referrer-policy": "strict-origin-when-cross-origin",
+        "x-ep-request-uuid": "3c78c409659702a5f42bde43000b14e3",
         "Cache-Control": "no-store, no-cache, private",
         "Pragma": "no-cache",
-        "x-runtime": "0.036847",
-        "Vary": "Origin",
-        "x-node": "bigweb38nuq",
-        "x-version-label": "easypost-202312122047-25e60d20e0-master",
+        "x-runtime": "0.186189",
+        "x-node": "bigweb43nuq",
+        "x-version-label": "easypost-202401041812-437974c716-master",
         "x-backend": "easypost",
-        "x-proxied": "intlb2nuq b3de2c47ef,extlb1nuq 003ad9bca0",
+        "x-canary": "direct",
+        "x-proxied": "intlb2nuq 2c48984abf,extlb2nuq 003ad9bca0",
         "Strict-Transport-Security": "max-age=31536000; includeSubDomains; preload"
       },
       "Status": {
