+++ EasyPost.Tests/cassettes/netstandard/user_service_with_parameters/all_children.json
@@ -1,7 +1,7 @@
 [
   {
-    "Duration": 298,
-    "RecordedAt": "2023-12-12T16:48:05.010043-07:00",
+    "Duration": 855,
+    "RecordedAt": "2024-01-04T12:10:29.008243-07:00",
     "Request": {
       "Body": "",
       "BodyContentType": "Text",
@@ -11,7 +11,7 @@
         "Authorization": "<REDACTED>",
         "User-Agent": "<REDACTED>"
       },
-      "Uri": "https://api.easypost.com/beta/users/children?page_size=5"
+      "Uri": "https://api.easypost.com/v2/users/children?page_size=5"
     },
     "Response": {
       "Body": "{\"children\":[],\"has_more\":false}",
@@ -29,15 +29,15 @@
         "x-download-options": "noopen",
         "x-permitted-cross-domain-policies": "none",
         "referrer-policy": "strict-origin-when-cross-origin",
-        "x-ep-request-uuid": "a51ac9626578f135e7897f7000535ab3",
+        "x-ep-request-uuid": "3c78c408659702a4f02289fc000b147c",
         "Cache-Control": "no-store, no-cache, private",
         "Pragma": "no-cache",
-        "x-runtime": "0.035569",
-        "Vary": "Origin",
-        "x-node": "bigweb33nuq",
-        "x-version-label": "easypost-202312122047-25e60d20e0-master",
+        "x-runtime": "0.150601",
+        "x-node": "bigweb43nuq",
+        "x-version-label": "easypost-202401041812-437974c716-master",
         "x-backend": "easypost",
-        "x-proxied": "intlb2nuq b3de2c47ef,extlb2nuq 003ad9bca0",
+        "x-canary": "direct",
+        "x-proxied": "intlb1nuq 2c48984abf,extlb2nuq 003ad9bca0",
         "Strict-Transport-Security": "max-age=31536000; includeSubDomains; preload"
       },
       "Status": {
